- Programado todos los CC - Solucionado bug en el que el cd de las habilidades solo llegaba hasta Integer.MAX_VALUE nanosegundos (Poco más de 2 segundos). - Habilidad 1 personaje casi programado. - Arreglado bug en el método onAttack. - Minions rebalanceado - Se ha modificado ligeramente la cámara para no estar tan cerca del personaje. - Corregido error: error #iterator() cannot be used nested.

+++ core/src/garcia/gonzalez/adrian/utiles/Habilidad.java
@@ -47,7 +47,7 @@ public class Habilidad {
 
         // Ponemos el tiempo actual y el tiempo que queda en nanosegundos para poderlo lanzar otra vez
         lastUsed = actual;
-        nextUse = actual + Math.round(cooldown * (1-cd));
+        nextUse = actual + Math.round(cooldown * ((double)(1-cd)));
 
     }
